+++ src/Umbraco.Web/Editors/Filters/MediaItemFormatter.cs
@@ -1,0 +1,45 @@
+using System;
+using AutoMapper;
+using Umbraco.Core;
+using Umbraco.Core.Models;
+using Umbraco.Web.Models.ContentEditing;
+
+namespace Umbraco.Web.Editors.Filters
+{
+    internal class MediaItemFormatter : ContentItemBaseFormatter<IMedia, MediaItemSave>
+    {
+        public MediaItemFormatter()
+            : this(Core.ApplicationContext.Current)
+        {
+        }
+
+        /// <summary>
+        /// Constructor
+        /// </summary>
+        /// <param name="applicationContext"></param>
+        public MediaItemFormatter(ApplicationContext applicationContext)
+            : base(applicationContext)
+        {
+        }
+
+        protected override IMedia GetExisting(MediaItemSave model)
+        {
+            return ApplicationContext.Services.MediaService.GetById(Convert.ToInt32(model.Id));
+        }
+
+        protected override IMedia CreateNew(MediaItemSave model)
+        {
+            var contentType = ApplicationContext.Services.ContentTypeService.GetMediaType(model.ContentTypeAlias);
+            if (contentType == null)
+            {
+                throw new InvalidOperationException("No content type found wth alias " + model.ContentTypeAlias);
+            }
+            return new Core.Models.Media(model.Name, model.ParentId, contentType);
+        }
+
+        protected override ContentItemDto<IMedia> MapFromPersisted(MediaItemSave model)
+        {
+            return Mapper.Map<IMedia, ContentItemDto<IMedia>>(model.PersistedContent);
+        }
+    }
+}
