+++ src/Umbraco.Web/Editors/MediaPostValidateAttribute.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Web.Http;
